Refactor Carousel (split on small components for code readability). Split TravelTypesIcons on two components to use separate mapper.
