+++ package.json
@@ -27,6 +27,7 @@
     "redux-logger": "^3.0.6",
     "sass": "^1.68.0",
     "slick-carousel": "^1.8.1",
+    "uuid": "^9.0.1",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
